+++ NEWS.md
@@ -6,6 +6,10 @@
 
 -   The alternate assignment operators `%<-%` and `%->%` collect result from {dtplyr} (or {dbplyr}) and convert it into the preferred data frame object. Also, the multiple or deconstructing assignment implemented in {zeallot} is also usable here in addition to the conversion into the preferred data frame.
 
+-   Row names were not printed when a data.frame is using {pillar}. It is now a column named and with class \<rownames\>.
+
+-   `as_matrix()` applied to a tbl_df object did not honor `.rownames`. Corrected.
+
 # svBase 0.9.0
 
 -   This is the first version of the package with main functions to manage data.frame/data.table/tibble objects.